update tests and gulp based on new file structure
