+++ test/unit/helpers-specs.js
@@ -1,46 +1,11 @@
 // transpile:mocha
 
-import chai from 'chai';
-import chaiAsPromised from 'chai-as-promised';
+import { generateMessage, packetFromBuffer } from '../../lib/helpers';
 import 'mochawait';
-import proxyquire from 'proxyquire';
 
 process.env.NODE_ENV = 'test';
 
-let usbStub = { '@noCallThru': true };
-let helpers = proxyquire('../../lib/helpers', {usb: usbStub });
-let generateMessage = helpers.generateMessage;
-let packetFromBuffer = helpers.packetFromBuffer;
-let getAdbInterface = helpers.getAdbInterface;
-let findAdbDevices = helpers.findAdbDevices;
-let selectBySerialNumber = helpers.selectBySerialNumber;
-
-import { ADB_COMMANDS, CONNECTION_TYPES, ADB_VALUES
-       , LIBUSB_VALUES } from '../../lib/constants';
-const LIBUSB_ENDPOINT_IN = LIBUSB_VALUES.LIBUSB_ENDPOINT_IN
-    , LIBUSB_ENDPOINT_OUT = LIBUSB_VALUES.LIBUSB_ENDPOINT_OUT;
-
-chai.should();
-let expect = chai.expect;
-chai.use(chaiAsPromised);
-
-// fake device setup
-let endpoints = [LIBUSB_ENDPOINT_IN, LIBUSB_ENDPOINT_OUT ];
-let deviceDescriptor = { idVendor: 0x04e8 // samsung
-                       , iSerialNumber: "12345" };
-let interfaceDescriptor = { bInterfaceClass: ADB_VALUES.ADB_CLASS
-                          , bInterfaceSubClass: ADB_VALUES.ADB_SUBCLASS
-                          , bInterfaceProtocol: ADB_VALUES.ADB_PROTOCOL };
-let iface = { descriptor: interfaceDescriptor
-            , endpoints: endpoints };
-let device = { interfaces: [iface,]
-             , deviceDescriptor: deviceDescriptor
-             , configDescriptor: {}
-             , open: () => { return "nothing"; } };
-let adbDevice = { device: device
-                , deviceInterface: iface
-                , serialNumber: "12345" };
-let deviceArray = [adbDevice];
+import { ADB_COMMANDS, CONNECTION_TYPES } from '../../lib/constants';
 
 describe('helper function tests', () => {
   describe('generateMessage tests', () => {
@@ -93,40 +58,6 @@ describe('helper function tests', () => {
       packetBuffer.writeUInt32LE(0, 12);
       let packet = packetFromBuffer(packetBuffer);
       (typeof packet.data).should.equal('undefined');
-    });
-  });
-  describe('getAdbInterface tests', () => {
-    it('should return an interface if there is one for ADB comms', () => {
-        getAdbInterface(device).should.not.be.null;
-    });
-    it('should return null if there are interfaces but no ADB interface', () => {
-      iface.descriptor.bInterfaceClass = 100;
-      expect(getAdbInterface(device)).to.be.a('null');
-    });
-    it('should return null if there are no interfaces at all', () => {
-      device.interfaces = null;
-      expect(getAdbInterface(device)).to.be.a('null');
-    });
-  });
-  describe('findAdbDevices tests', () => {
-    usbStub.getDeviceList = ()=> { return [device]; };
-    it('should return an array with a length of zero', () => {
-      expect(findAdbDevices()).to.be.empty;
-    });
-    it('should return an object if there was a device with an adb interface', () => {
-      device.interfaces = [iface];
-      iface.descriptor.bInterfaceClass = 255;
-      expect(findAdbDevices()).to.not.be.empty;
-    });
-  });
-  describe('selectBySerialNumber tests', () => {
-    it('should return a device if a device with that serial number is available', () => {
-      expect(selectBySerialNumber(deviceArray, "12345")).should.not.be.null;
-    });
-    it('should throw an error if there is no device with that serial number', () => {
-      () => {
-        selectBySerialNumber(deviceArray, "54321");
-      }.should.throw();
     });
   });
 });
